requirements.txt
fastapi
uvicorn
python-multipart
langgraph
langchain-google-genai
langchain
python-dotenv
pymupdf
pillow
pydantic
google[generativeai]